Keep intermediate resolved addresses
Currently when resolving addresses, only the original and finaly IP
address are passed to the next stages
In Wasm, when dialing WSS, the hostname is required to make a
successful connection.  When resolving /dnsaddr addresses, these
addresses are intermediate addresses.  Keeping them allows /dnsaddr
WSS addresses on  WASM to succeed.  The intermediate addresses will
usually be filtered out before dialing